refactor: comment away all new code
